+++ examples/get-table.ts
@@ -1,21 +1,24 @@
-import { FlightClient, listFlights, doGetTable } from '../src';
+import { FlightClient } from '../src';
 
 async function main() {
   const client = new FlightClient('localhost:8815');
-  const flights = await listFlights(client);
-  
-  const flight = flights[0];
-  const ticket = flight.endpoint?.[0].ticket?.ticket;
 
-  if (!ticket) {
-    throw new Error('No ticket found');
+  try {
+    for await (const flight of client.listFlights()) {
+      const ticket = flight.endpoints[0]?.ticket;
+
+      if (ticket) {
+        const table = await client.getTable(ticket);
+        console.log(table.toString());
+        return;
+      }
+    }
+
+    throw new Error('No Flight endpoint with a ticket was found');
   }
-
-  const table = await doGetTable(client, ticket);
-
-  console.log(table.toString());
-
-  await client.close();
+  finally {
+    await client.close();
+  }
 }
 
 main().catch(console.error);