+++ examples/list-flights.ts
@@ -1,16 +1,18 @@
-import { FlightClient, listFlights } from '../src';
+import { FlightClient } from '../src';
 
 async function main() {
   const client = new FlightClient('localhost:8815');
-  const flights = await listFlights(client);
 
-  console.log('Available flights:');
+  try {
+    console.log('Available flights:');
 
-  for (const flight of flights) {
-    console.log('-', flight.flightDescriptor?.path);
+    for await (const flight of client.listFlights()) {
+      console.log('-', flight.descriptor);
+    }
   }
-
-  await client.close();
+  finally {
+    await client.close();
+  }
 }
 
 main().catch(console.error);