+++ examples/put-table.ts
@@ -1,4 +1,4 @@
-import { FlightClient, doPutTable } from '../src';
+import { FlightClient, pathDescriptor } from '../src';
 import { tableFromArrays } from 'apache-arrow';
 
 async function main() {
@@ -7,13 +7,21 @@ async function main() {
   const table = tableFromArrays({
     id: [4, 5, 6],
     name: ['Dave', 'Eve', 'Frank']
-  })
+  });
 
-  await doPutTable(client, table, ['uploaded', 'table']);
+  try {
+    for await (const result of client.doPut(
+      pathDescriptor('uploaded', 'table'),
+      table
+    )) {
+      console.log('Server metadata:', result.appMetadata);
+    }
 
-  console.log('Table uploaded');
-
-  await client.close();
+    console.log('Table uploaded');
+  }
+  finally {
+    await client.close();
+  }
 }
 
 main().catch(console.error);